parseColorsStream: fix empty/white lines detection
When input line is empty, sscanf does not write a '\0' to the %s target.
Initialize its first byte to '\0' to properly detect empty lines after.

Next, we do not want to stop reading the file upon reading an empty line:
use continue instead of break.

The latter was not noticed because of the former.